requirements.txt
click==8.1.7
cvzone==1.6.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
ImageHash==4.3.1
numpy
opencv-python==4.10.0.84
pillow==10.4.0
PyWavelets==1.7.0
scipy==1.14.1
setuptools==75.1.0
wheel==0.44.0
